git update-ref -d MERGE_HEAD

+++ Source/ACE.Database/Models/Shard/CharacterLogin.cs
@@ -9,7 +9,7 @@ namespace ACE.Database.Models.Shard
         public uint AccountId { get; set; }
         public string AccountName { get; set; }
         public string SessionIP { get; set; }
-        public uint CharacterId { get; set; }
+        public ulong CharacterId { get; set; }
         public string CharacterName { get; set; }
         public DateTime LoginDateTime { get; set; }
